+++ R/rbackground.R
@@ -66,7 +66,7 @@ print.RPolarsRThreadHandle = function(x, ...) as.character(x) |> cat("\n")
 #'   Sys.sleep(.1)
 #'   x * 0.43
 #' }, in_background = TRUE)$alias("kml")
-#' handle = pl$LazyFrame(mtcars)$with_columns(prexpr)$collect_in_background()
+#' handle = as_polars_lf(mtcars)$with_columns(prexpr)$collect_in_background()
 #' if (!handle$is_finished()) print("not done yet")
 #' df = handle$join() # get result
 #' df