+++ tests/testthat/test-enable_string_cache.R
@@ -19,8 +19,8 @@ test_that("string cache", {
   before = pl$using_string_cache()
   value = pl$with_string_cache({
     during = pl$using_string_cache()
-    df1 = pl$DataFrame(head(iris, 2))
-    df2 = pl$DataFrame(tail(iris, 2))
+    df1 = as_polars_df(head(iris, 2))
+    df2 = as_polars_df(tail(iris, 2))
     pl$concat(list(df1, df2)) # should not error
   })
   after = pl$using_string_cache()
@@ -33,8 +33,8 @@ test_that("string cache", {
   before = pl$using_string_cache()
   value = pl$with_string_cache({
     during = pl$using_string_cache()
-    df1 = pl$DataFrame(head(iris, 2))
-    df2 = pl$DataFrame(tail(iris, 2))
+    df1 = as_polars_df(head(iris, 2))
+    df2 = as_polars_df(tail(iris, 2))
     pl$concat(list(df1, df2)) # should not error
   })
   after = pl$using_string_cache()